+++ CHANGELOG.md
@@ -8,8 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/) and this 
 ### Changed
 - Updated dependencies to the latest versions.
 - The bastion's public key will be created from the private key submitted in the `Settings` page. ([#318][i318])
+- The unused temporary dir has been removed from the `Settings` page. ([#165][i165]) 
 
 [i318]: https://github.com/pacoorozco/ssham/issues/318
+[i165]: https://github.com/pacoorozco/ssham/issues/165
 
 ## 0.15.1 - 2022-05-27